==> code/notebooks/2-CognitiveData.ipynb <==
Author: Kevin ALBERT

Created: Feb 2021

# !pip install pyarrow fastparquet

# current python SDK versions:
!pip freeze | grep azure-ai-textanalytics

Generated API keys and endpoints

azure_textanalytics_key = '2180035c81284813a4d457640c87f017'
azure_textanalytics_endpoint = 'https://westeurope.api.cognitive.microsoft.com/'
azure_textanalytics_url = azure_textanalytics_endpoint + 'text/analytics/v3.0'

import warnings
warnings.filterwarnings("ignore")
import os
import requests
import json
import time
import uuid
import glob
from io import BytesIO
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib.patches as patches
%matplotlib inline

from azure.ai.textanalytics import TextAnalyticsClient         # for text analytics
from azure.core.credentials import AzureKeyCredential          # for text analytics

API docs: https://westeurope.dev.cognitive.microsoft.com/docs/services/TextAnalytics-v3-0/operations/Languages

client = TextAnalyticsClient(azure_textanalytics_endpoint, AzureKeyCredential(azure_textanalytics_key))

## key_phrases

# documents = ["My cat might need to see a veterinarian."]
documents = ["I had a wonderful trip to Seattle last week.",
             "This was a waste of my time. The speaker put me to sleep.",
             "I don't feel that my skills and capacities are taking into account."]

keyphrases_response = client.extract_key_phrases(documents)
[i.__dict__ for i in keyphrases_response]

for idx, doc in enumerate(keyphrases_response):
    print(idx, doc)

# showing the last line, this is a list
doc.key_phrases

## sentiment

documents = ["Dit is een heel mooie show.",
             "This was a waste of my time. The speaker put me to sleep."]
# documents = ["I had a wonderful trip to Seattle last week.",
#              "This was a waste of my time. The speaker put me to sleep.",
#              "I don't feel that my skills and capacities are taking into account."]

sentiment_response = client.analyze_sentiment(documents)
[i.__dict__ for i in sentiment_response]

for idx, doc in enumerate(sentiment_response):
    print(idx, doc)

doc.sentiment

doc.sentences[0]

## named_entities

documents = ["I had a wonderful trip to Seattle last week.",
             "This was a waste of my time. The speaker put me to sleep.",
             "I don't feel that my skills and capacities are taking into account."]

response = client.recognize_entities(documents)
[i.__dict__ for i in response]

save to *.csv

## load data

We need to run this for 3 columns ! (repeat this process 3x)
 1. Any_Additional_Suggestion_To_Improve
 1. What_Felt_Best_During_This_Last_Month  
 1. What_Should_We_Do_To_Make_You_Feel_Better_And_Increase_Your_Overall_Satisfaction
 
don't forget to remove null values, remove old csv files and rename the final datasets 

synthetic_df = pd.read_csv("../../data/bronze/sdworxHRhackathon2021.csv")

synthetic_df.head(2).T

# here you replace the columns 3x as well:
df = pd.DataFrame({'col1': synthetic_df["What_Should_We_Do_To_Make_You_Feel_Better_And_Increase_Your_Overall_Satisfaction"].values})
# replace nan
df['col1'] = df['col1'].fillna('none')
df

#### generating entities

# replace 'key' and 'endpoint':
azure_textanalytics_key = '2180035c81284813a4d457640c87f017'
azure_textanalytics_endpoint = 'https://westeurope.api.cognitive.microsoft.com/'

import pandas as pd
from azure.core.credentials import AzureKeyCredential
from azure.ai.textanalytics import TextAnalyticsClient
client = TextAnalyticsClient(azure_textanalytics_endpoint, AzureKeyCredential(azure_textanalytics_key))

result = pd.DataFrame()
temp   = []
start  = 0      # index id start (ex: 0)
stop   = 100   # index id stop  (ex: 5000)
step   = 1      # max 5 (ex: 5)
column = "col1" # from dataframe
# https://docs.microsoft.com/en-us/azure/cognitive-services/text-analytics/concepts/data-limits

for i in range(start, stop, step): # 0-5000, 5000-10000, etc... (~10min)
    documents = (df[column].iloc[i:i+step]).tolist()
    response = client.recognize_entities(documents)
    print(i, end='\r') # counter
    for idx, doc in enumerate(response):
        comment = pd.DataFrame()
        for idy, entity in enumerate(doc.entities):
            comment.loc[idy, "document"] = str(documents[idx])
            comment.loc[idy, "entities"] = str(entity)
            comment.loc[idy, "entity_text"] = str(entity.text)
            comment.loc[idy, "entity_category"] = str(entity.category)
            comment.loc[idy, "entity_subcategory"] = str(entity.subcategory)
            comment.loc[idy, "entity_confidence_score"] = str(entity.confidence_score)
        temp.append(comment)
        
result = pd.concat(temp, axis=0, ignore_index=True)
result.to_csv("entities_"+str(start)+"_"+str(stop)+".csv", index=False)
result

# Merge it all together into one csv
import glob

all_files = glob.glob("entities_*.csv")
li = []
for filename in all_files:
    dataset = pd.read_csv(filename)
    li.append(dataset)
result = pd.concat(li, axis=0, ignore_index=True)
result.to_parquet("../../data/bronze/entities_What_Should_We_Do_To_Make_You_Feel_Better_And_Increase_Your_Overall_Satisfaction.parquet")
result

#### generating sentiment

# replace 'key' and 'endpoint':
azure_textanalytics_key = '2180035c81284813a4d457640c87f017'
azure_textanalytics_endpoint = 'https://westeurope.api.cognitive.microsoft.com/'

import pandas as pd
from azure.core.credentials import AzureKeyCredential
from azure.ai.textanalytics import TextAnalyticsClient
client = TextAnalyticsClient(azure_textanalytics_endpoint, AzureKeyCredential(azure_textanalytics_key))

result = pd.DataFrame()
temp   = []
start  = 0      # index id start (ex: 0)
stop   = 100   # index id stop  (ex: 5000)
step   = 1      # max 5 (ex: 5)
column = "col1" # from dataframe
# https://docs.microsoft.com/en-us/azure/cognitive-services/text-analytics/concepts/data-limits

for i in range(start, stop, step): # 0-5000, 5000-10000, etc... (~10min)
    documents = (df[column].iloc[i:i+step]).tolist()

    sentiment_response = client.analyze_sentiment(documents)
    print(i, end='\r') # second counter
    for idx, doc in enumerate(sentiment_response):
        comment = pd.DataFrame()
        comment.loc[idy, "document"] = str(documents[idx])
        comment.loc[idy, "sentiment"] = str(doc.sentiment)
    temp.append(comment)

result = pd.concat(temp, axis=0, ignore_index=True)
result.to_csv("sentiment_"+str(start)+"_"+str(stop)+".csv", index=False)
result

# Merge it all together into one csv
import glob

all_files = glob.glob("sentiment_*.csv")
li = []
for filename in all_files:
    dataset = pd.read_csv(filename)
    li.append(dataset)
result = pd.concat(li, axis=0, ignore_index=True)
result.to_parquet("../../data/bronze/sentiment_What_Should_We_Do_To_Make_You_Feel_Better_And_Increase_Your_Overall_Satisfaction.parquet", index=False)
result

#### generating keyphrases

# replace 'key' and 'endpoint':
azure_textanalytics_key = '2180035c81284813a4d457640c87f017'
azure_textanalytics_endpoint = 'https://westeurope.api.cognitive.microsoft.com/'

import pandas as pd
from azure.core.credentials import AzureKeyCredential
from azure.ai.textanalytics import TextAnalyticsClient
client = TextAnalyticsClient(azure_textanalytics_endpoint, AzureKeyCredential(azure_textanalytics_key))

result = pd.DataFrame()
temp   = []
start  = 0      # index id start (ex: 0)
stop   = 100   # index id stop  (ex: 5000)
step   = 1      # max 5 (ex: 5)
column = "col1" # from dataframe
# https://docs.microsoft.com/en-us/azure/cognitive-services/text-analytics/concepts/data-limits

for i in range(start, stop, step): # 0-5000, 5000-10000, etc... (~10min)
    documents = (df[column].iloc[i:i+step]).tolist()

    keyphrases_response = client.extract_key_phrases(documents)
    print(i, end='\r') # second counter
    for idx, doc in enumerate(keyphrases_response):
        comment = pd.DataFrame()
        comment.loc[idy, "document"] = str(documents[idx])
        comment.loc[idy, "key_phrases"] = str(doc.key_phrases)
    temp.append(comment)

result = pd.concat(temp, axis=0, ignore_index=True)
result.to_csv("key_phrases_"+str(start)+"_"+str(stop)+".csv", index=False)
result

doc.key_phrases

# Merge it all together into one csv
import glob

all_files = glob.glob("key_phrases_*.csv")
li = []
for filename in all_files:
    dataset = pd.read_csv(filename)
    li.append(dataset)
result = pd.concat(li, axis=0, ignore_index=True)
result.to_parquet("../../data/bronze/key_phrases_What_Should_We_Do_To_Make_You_Feel_Better_And_Increase_Your_Overall_Satisfaction.parquet", index=False)
result

# we now have saved all the files we need
!ls -al ../../data/bronze/
