==> code/notebooks/3-DataGraph_Suggestion.ipynb <==
Author: Kevin ALBERT

Created: Feb 2021

from py2neo import Graph, Node, Relationship
import pandas as pd
from IPython.display import Javascript
from fuzzywuzzy import process, fuzz

import platform
import psutil
import os

conda_version = ! conda -V
print(f"conda : {conda_version[0].split()[1]}")
pip_version = ! pip -V
print(f"pip   : {pip_version[0].split()[1]}")
python_version = ! python -V
print(f"python: {python_version[0].split()[1]}")
pandas_version = ! pip list |grep -i pandas
print(f"pandas: {pandas_version[0].split()[1]}")
py2neo_version = ! pip list |grep -i py2neo
print(f"py2neo: {py2neo_version[0].split()[1]}")

server  = "168.63.44.210"
port    = "7687"
user    = "neo4j"
passw   = "digityser"
db_name = "neo4j"        # default name (v4.x)

# load graph connection instance
graph = Graph(host=server, auth=(user, passw), name=db_name, encrypted=False)

## prepare data
We may be trying to do it for all 3 dataset groups !

# import the datasets
synthetic_df = pd.read_parquet("../../data/silver/synthetic_data_processed.parquet")
entities_df = pd.read_parquet("../../data/bronze/entities_Any_Additional_Suggestion_To_Improve.parquet")
sentiment_df = pd.read_parquet("../../data/bronze/sentiment_Any_Additional_Suggestion_To_Improve.parquet")
keyphrase_df = pd.read_parquet("../../data/bronze/key_phrases_Any_Additional_Suggestion_To_Improve.parquet")

# creating unique patient ID
synthetic_df["id"] = synthetic_df.index

# replace nan
synthetic_df['What_Felt_Best_During_This_Last_Month'] = synthetic_df['What_Felt_Best_During_This_Last_Month'].fillna('No Feeling')
synthetic_df['What_Should_We_Do_To_Make_You_Feel_Better_And_Increase_Your_Overall_Satisfaction'] = synthetic_df['What_Should_We_Do_To_Make_You_Feel_Better_And_Increase_Your_Overall_Satisfaction'].fillna('No Satisfaction')
synthetic_df['Any_Additional_Suggestion_To_Improve'] = synthetic_df['Any_Additional_Suggestion_To_Improve'].fillna('No Improvement')

synthetic_df

# count missing values, sorted (high missing > 5%)
synthetic_df.isnull().apply(lambda x: x.sum() * 100 / len(synthetic_df)).round(1).sort_values(ascending=True)

# save file to database /import
file_synthetic = "synthetic_data_processed.csv"
synthetic_df.to_csv("../../neo4j/import/"+file_synthetic, sep=',', index=False)

file_entities1 = "entities1.csv"
entities_df.to_csv("../../neo4j/import/"+file_entities1, sep=',', index=False)

file_sentiment1 = "sentiment1.csv"
sentiment_df.to_csv("../../neo4j/import/"+file_sentiment1, sep=',', index=False)

file_keyphrase1 = "keyphrases1.csv"
keyphrase_df.to_csv("../../neo4j/import/"+file_keyphrase1, sep=',', index=False)

## import data in DB

[link to arrows data modeling tool](https://arrows.app/#/import/json=eyJncmFwaCI6eyJzdHlsZSI6eyJub2RlLWNvbG9yIjoiIzRDOEVEQSIsImJvcmRlci13aWR0aCI6MCwiYm9yZGVyLWNvbG9yIjoiIzAwMDAwMCIsInJhZGl1cyI6NzUsIm5vZGUtcGFkZGluZyI6NSwib3V0c2lkZS1wb3NpdGlvbiI6ImF1dG8iLCJjYXB0aW9uLXBvc2l0aW9uIjoiaW5zaWRlIiwiY2FwdGlvbi1tYXgtd2lkdGgiOjIwMCwiY2FwdGlvbi1jb2xvciI6IiNmZmZmZmYiLCJjYXB0aW9uLWZvbnQtc2l6ZSI6MjAsImNhcHRpb24tZm9udC13ZWlnaHQiOiJub3JtYWwiLCJsYWJlbC1wb3NpdGlvbiI6Imluc2lkZSIsImxhYmVsLWNvbG9yIjoiIzAwMDAwMCIsImxhYmVsLWJhY2tncm91bmQtY29sb3IiOiIjZmZmZmZmIiwibGFiZWwtYm9yZGVyLWNvbG9yIjoiIzg0ODQ4NCIsImxhYmVsLWJvcmRlci13aWR0aCI6MywibGFiZWwtZm9udC1zaXplIjoyMCwibGFiZWwtcGFkZGluZyI6NSwibGFiZWwtbWFyZ2luIjo0LCJkaXJlY3Rpb25hbGl0eSI6ImRpcmVjdGVkIiwiZGV0YWlsLXBvc2l0aW9uIjoiYWJvdmUiLCJkZXRhaWwtb3JpZW50YXRpb24iOiJwYXJhbGxlbCIsImFycm93LXdpZHRoIjozLCJhcnJvdy1jb2xvciI6IiM4NDg0ODQiLCJtYXJnaW4tc3RhcnQiOjUsIm1hcmdpbi1lbmQiOjUsIm1hcmdpbi1wZWVyIjoyMCwiYXR0YWNobWVudC1zdGFydCI6Im5vcm1hbCIsImF0dGFjaG1lbnQtZW5kIjoibm9ybWFsIiwidHlwZS1jb2xvciI6IiM4NDg0ODQiLCJ0eXBlLWJhY2tncm91bmQtY29sb3IiOiIjZmZmZmZmIiwidHlwZS1ib3JkZXItY29sb3IiOiIjODQ4NDg0IiwidHlwZS1ib3JkZXItd2lkdGgiOjAsInR5cGUtZm9udC1zaXplIjoyMSwidHlwZS1wYWRkaW5nIjo1LCJwcm9wZXJ0eS1wb3NpdGlvbiI6Im91dHNpZGUiLCJwcm9wZXJ0eS1jb2xvciI6IiM4NDg0ODQiLCJwcm9wZXJ0eS1mb250LXNpemUiOjIwLCJwcm9wZXJ0eS1mb250LXdlaWdodCI6Im5vcm1hbCJ9LCJub2RlcyI6W3siaWQiOiJuMCIsInBvc2l0aW9uIjp7IngiOjAsInkiOjB9LCJjYXB0aW9uIjoiRW1wbG95ZWUiLCJzdHlsZSI6e30sImxhYmVscyI6W10sInByb3BlcnRpZXMiOnt9fV0sInJlbGF0aW9uc2hpcHMiOltdfSwiZGlhZ3JhbU5hbWUiOiJzZHdvcnhIUmhhY2thdGhvbjIwMjFfZGF0YU1vZGVsIn0=)

graph.run("CREATE CONSTRAINT ON (e:Employee) ASSERT e.id IS UNIQUE")

graph.run("CREATE CONSTRAINT ON (a:Assignmentscore) ASSERT a.value IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_synthetic+"""' AS line FIELDTERMINATOR ','
MERGE (e:Employee {id:toInteger(line.id), fullName:(line.full_name), email:(line.email)})
MERGE (a:Assignmentscore {value:toInteger(line.how_was_your_last_month_in_your_assignment)})
MERGE (e)-[:GAVE_ASSIGNMENT_SCORE]->(a)
"""
display(graph.run(query).stats())

graph.run("CREATE CONSTRAINT ON (d:Departmentscore) ASSERT d.value IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_synthetic+"""' AS line FIELDTERMINATOR ','
MATCH (e:Employee {id:toInteger(line.id)})
MERGE (d:Departmentscore {value:toInteger(line.How_Was_Your_Last_Month_Within_Your_Department)})
MERGE (e)-[:GAVE_DEPARTMENT_SCORE]->(d)
"""
display(graph.run(query).stats())

graph.run("CREATE CONSTRAINT ON (c:Companyscore) ASSERT c.value IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_synthetic+"""' AS line FIELDTERMINATOR ','
MATCH (e:Employee {id:toInteger(line.id)})
MERGE (c:Companyscore {value:toInteger(line.How_Was_Your_Last_Month_With_Us)})
MERGE (e)-[:GAVE_COMPANY_SCORE]->(c)
"""
display(graph.run(query).stats())

#### feeling

graph.run("CREATE CONSTRAINT ON (f:Feeling) ASSERT f.description IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_synthetic+"""' AS line FIELDTERMINATOR ','
MATCH (e:Employee {id:toInteger(line.id)})
MERGE (f:Feeling {description:(line.What_Felt_Best_During_This_Last_Month)})
MERGE (e)-[:GAVE_BEST_FEELING]->(f)
"""
display(graph.run(query).stats())

#### Satisfaction

graph.run("CREATE CONSTRAINT ON (s:Satisfaction) ASSERT s.description IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_synthetic+"""' AS line FIELDTERMINATOR ','
MATCH (e:Employee {id:toInteger(line.id)})
MERGE (s:Satisfaction {description:(line.What_Should_We_Do_To_Make_You_Feel_Better_And_Increase_Your_Overall_Satisfaction)})
MERGE (e)-[:GAVE_SATISFACTION_IMPROVEMENTS]->(s)
"""
display(graph.run(query).stats())

#### Suggestion

graph.run("CREATE CONSTRAINT ON (s:Suggestion) ASSERT s.description IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_synthetic+"""' AS line FIELDTERMINATOR ','
MATCH (e:Employee {id:toInteger(line.id)})
MERGE (s:Suggestion {description:(line.Any_Additional_Suggestion_To_Improve)})
MERGE (e)-[:GAVE_IMPROVEMENT_SUGGESTIONS]->(s)
"""
display(graph.run(query).stats())

#### cognitive data

graph.run("CREATE CONSTRAINT ON (e:Event) ASSERT e.value IS UNIQUE")
graph.run("CREATE CONSTRAINT ON (l:Location) ASSERT l.value IS UNIQUE")
graph.run("CREATE CONSTRAINT ON (p:Product) ASSERT p.value IS UNIQUE")
graph.run("CREATE CONSTRAINT ON (s:Skill) ASSERT s.value IS UNIQUE")
graph.run("CREATE CONSTRAINT ON (q:Quantity) ASSERT q.value IS UNIQUE")
graph.run("CREATE CONSTRAINT ON (p:PersonType) ASSERT p.value IS UNIQUE")

entities_df

# count missing values, sorted (high missing > 5%)
entities_df.isnull().apply(lambda x: x.sum() * 100 / len(entities_df)).round(1).sort_values(ascending=True)

#### replace Suggestion Node to either Feeling, Satisfaction or Suggestion into the other 2 notebooks!!
here it hould be Suggestion node

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_entities1+"""' AS line FIELDTERMINATOR ','
WITH line WHERE line.entity_category = 'Event'
MATCH (suggestion:Suggestion {description:(line.document)})
MERGE (event:Event {value:(line.entity_text)})
MERGE (event)-[:EVENT_IN]->(suggestion)
"""
display(graph.run(query).stats())
query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_entities1+"""' AS line FIELDTERMINATOR ','
WITH line WHERE line.entity_category = 'Location'
MATCH (suggestion:Suggestion {description:(line.document)})
MERGE (location:Location {value:(line.entity_text)})
MERGE (location)-[:LOCATION_IN]->(suggestion)
"""
display(graph.run(query).stats())
query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_entities1+"""' AS line FIELDTERMINATOR ','
WITH line WHERE line.entity_category = 'Product'
MATCH (suggestion:Suggestion {description:(line.document)})
MERGE (product:Product {value:(line.entity_text)})
MERGE (product)-[:PRODUCT_IN]->(suggestion)
"""
display(graph.run(query).stats())
query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_entities1+"""' AS line FIELDTERMINATOR ','
WITH line WHERE line.entity_category = 'Skill'
MATCH (suggestion:Suggestion {description:(line.document)})
MERGE (skill:Skill {value:(line.entity_text)})
MERGE (skill)-[:SKILL_IN]->(suggestion)
"""
display(graph.run(query).stats())
query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_entities1+"""' AS line FIELDTERMINATOR ','
WITH line WHERE line.entity_category = 'Quantity'
MATCH (suggestion:Suggestion {description:(line.document)})
MERGE (quantity:Quantity {value:(line.entity_text)})
MERGE (quantity)-[:QUANTITY_IN]->(suggestion)
"""
display(graph.run(query).stats())
query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_entities1+"""' AS line FIELDTERMINATOR ','
WITH line WHERE line.entity_category = 'PersonType'
MATCH (suggestion:Suggestion {description:(line.document)})
MERGE (persontype:PersonType {value:(line.entity_text)})
MERGE (persontype)-[:PERSONTYPE_IN]->(suggestion)
"""
display(graph.run(query).stats())

#### sentiment

graph.run("CREATE CONSTRAINT ON (s:Sentiment) ASSERT s.value IS UNIQUE")

sentiment_df

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file_sentiment1+"""' AS line FIELDTERMINATOR ','
MATCH (suggestion:Suggestion {description:(line.document)})
MERGE (sentiment:Sentiment {value:(line.sentiment)})
MERGE (sentiment)-[:SENTIMENT_IN]->(suggestion)
"""
display(graph.run(query).stats())

## Querying

%%time
query = """
MATCH (s:Sentiment)-[:SENTIMENT_IN]->()<-[rel]-(e:Employee)
WITH e, rel, COLLECT(s.value) AS sentiment
RETURN e.email, count(sentiment) AS sentiment_amount
ORDER BY sentiment_amount DESC
"""
graph.run(query).to_data_frame()

query = """
MATCH (s:Sentiment)-[rel]-()
RETURN count(s) AS total
"""
graph.run(query).to_data_frame()['total']

MATCH (s:Sentiment)-[rel]-()
RETURN count(s)

## reset DB

# delete database and restart (wait ~2min)
import os
os.system(" cd ../.. && \
            sudo docker-compose down && \
            sudo rm -Rf neo4j/data/databases/neo4j && \
            sudo rm -Rf neo4j/data/transactions/neo4j && \
            sudo docker-compose up --build &")
