==> code/notebooks/6-DataGraph_Querying.ipynb <==
from py2neo import Graph, Node, Relationship
import pandas as pd
from IPython.display import Javascript
from fuzzywuzzy import process, fuzz

import platform
import psutil
import os

server  = "168.63.44.210"
port    = "7687"
user    = "neo4j"
passw   = "digityser"
db_name = "neo4j"        # default name (v4.x)

# load graph connection instance
graph = Graph(host=server, auth=(user, passw), name=db_name, encrypted=False)

#### detected nodes that have no value

query = """
MATCH (s)
WHERE s.description STARTS WITH 'No '
RETURN s.description AS data
"""
graph.run(query).to_data_frame()

query = """
MATCH (s)
WHERE s.description STARTS WITH 'No '
DETACH DELETE s
"""
display(graph.run(query).stats())

## what percentages has a positive, negative and neutral sentiment ?

%%time
query = """
MATCH (s:Sentiment {value:'positive'})-[:SENTIMENT_IN]->()<-[rel]-(e:Employee)
WITH count(s) AS totalPositive
MATCH (s:Sentiment {value:'negative'})-[:SENTIMENT_IN]->()<-[rel]-(e:Employee)
WITH totalPositive, count(s) AS totalNegative
MATCH (s:Sentiment {value:'neutral'})-[:SENTIMENT_IN]->()<-[rel]-(e:Employee)
WITH totalPositive, totalNegative, count(s) AS totalNeutral
MATCH (s:Sentiment)-[:SENTIMENT_IN]->()<-[rel]-(e:Employee)
WITH totalPositive, totalNegative, totalNeutral, count(s) AS total
RETURN toFloat((totalPositive)/toFloat(total))*100 AS pos,
toFloat((totalNegative)/toFloat(total))*100 AS neg,
toFloat((totalNeutral)/toFloat(total))*100 AS neu
"""
graph.run(query).to_data_frame()

#### List all the entities related to positive, negative and neutral sentiment

# neutral entities:
query = """
MATCH (s:Sentiment {value:'positive'})-[:SENTIMENT_IN]->(feeling:Feeling)
WITH feeling
MATCH (s:Sentiment {value:'positive'})-[:SENTIMENT_IN]->(satisfaction:Satisfaction)
WITH feeling, satisfaction
MATCH (s:Sentiment {value:'positive'})-[:SENTIMENT_IN]->(suggestion:Suggestion)
WITH feeling, satisfaction, suggestion
MATCH (satisfaction)<-[r:SKILL_IN]-(d)
RETURN DISTINCT d.value AS positive_entities
"""
graph.run(query).to_data_frame()

# negative entities:
query = """
MATCH (s:Sentiment {value:'negative'})-[:SENTIMENT_IN]->(feeling:Feeling)
WITH feeling
MATCH (s:Sentiment {value:'negative'})-[:SENTIMENT_IN]->(satisfaction:Satisfaction)
WITH feeling, satisfaction
MATCH (s:Sentiment {value:'negative'})-[:SENTIMENT_IN]->(suggestion:Suggestion)
WITH feeling, satisfaction, suggestion
MATCH (satisfaction)<-[r:SKILL_IN]-(d)
RETURN DISTINCT d.value AS negative_entities
"""
graph.run(query).to_data_frame()

# neutral entities:
query = """
MATCH (s:Sentiment {value:'neutral'})-[:SENTIMENT_IN]->(feeling:Feeling)
WITH feeling
MATCH (s:Sentiment {value:'neutral'})-[:SENTIMENT_IN]->(satisfaction:Satisfaction)
WITH feeling, satisfaction
MATCH (s:Sentiment {value:'neutral'})-[:SENTIMENT_IN]->(suggestion:Suggestion)
WITH feeling, satisfaction, suggestion
MATCH (satisfaction)<-[r:SKILL_IN]-(d)
RETURN DISTINCT d.value AS neutral_entities
"""
graph.run(query).to_data_frame()
