==> code/notebooks/7-DataSet_PrepforML.ipynb <==
Author: Kevin ALBERT

Created: Feb 2021

import pandas as pd

# pd.describe_option('display')            # show all pandas options, parameters can slow down notebook
pd.set_option('display.max_colwidth', 100) # default 50, the maximum width in characters of a column
pd.set_option('display.max_columns', 40)   # default 20, the maximum amount of columns in view 
pd.set_option('display.max_rows', 60)      # default 60, the maximum amount of rows in view

!conda -V
!python -V
!conda list |grep pandas

## load data

synthetic_df = pd.read_csv("../../neo4j/import/synthetic_data_processed.csv")
sentiment1_df = pd.read_csv("../../neo4j/import/sentiment1.csv")
sentiment2_df = pd.read_csv("../../neo4j/import/sentiment2.csv")
sentiment3_df = pd.read_csv("../../neo4j/import/sentiment3.csv")

## summary

# concise summary (shape, memory use, data types, nan's)
synthetic_df.info()

## data checks

# count distinct (=unique) observations (+ missing), sorted (high cardinality > 390)
synthetic_df.nunique(dropna=False).sort_values(ascending=False)

# count missing values, sorted (high missing > 5%)
synthetic_df.isnull().apply(lambda x: x.sum() * 100 / len(synthetic_df)).round(1).sort_values(ascending=True)

synthetic_df

sentiment1_df

sentiment2_df

sentiment3_df

## data cleaning

import sys  
sys.path.insert(0, '../scripts')
import blobstorage
from io import BytesIO

#### The idea is to create a prediction feature, based on the 'sentiment' of the text

synthetic_df['improve_sentiment'] = sentiment1_df["sentiment"]
synthetic_df['feeling_sentiment'] = sentiment2_df["sentiment"]
synthetic_df['satisfaction_sentiment'] = sentiment3_df["sentiment"]

## save new dataset

# store to cloud datalake as *.parquet (preserve schema)
blobstorage.write_blob("silver/synthetic_data_processed.parquet", synthetic_df)

# store local as *.csv files
synthetic_df.to_csv("../../data/silver/synthetic_data_processed.csv", sep=',', index=False)
# store local as *.parquet
synthetic_df.to_parquet("../../data/silver/synthetic_data_processed.parquet")
